Refatoração: organiza código em controller, service, repository
