+++ src/services/redeemedRewardsServices/readRedeemedRewards.service.ts
@@ -1,0 +1,15 @@
+import { RedeemedRewardsRepository } from "../../repositories/redeemedRewardsRepositories/redeemedRewards.repository.js";
+
+export class RedeemedRewardsService {
+  private repository = new RedeemedRewardsRepository();
+
+  async getUserRedeemedRewards(userId: string) {
+    const data = await this.repository.findByUserId(userId);
+
+    if (!data || data.length === 0) {
+      return null;
+    }
+
+    return data;
+  }
+}
